requirements.txt
tensorflow
numpy
pandas
matplotlib
scikit-learn
opencv-python
pyyaml